0.027
removed bad smells, added new item: arrow — deflects ball based on
arrow’s current direction

+++ Tau/powerup.h
@@ -38,7 +38,7 @@ protected:
 private:
     QTimer* disabler;
     Player* affectedPlayer;
-    double angle = 0;
+    int angle = 0; // in degrees
     void keepRotating();
     double opacity = 0;
     void fadeIn();
